fix: seperate training case
